Adds logic to update selected cell
Updates events.js with logic to only update a cell if it has not already
been claimed by a player. If the cell has been claimed, it will be noted
in the log and the player can try again. Log message will be replaced
with UI message in a future commit.