Adds gameOver logic
Updates isGameOver function in logic.js to know when the game is over.
Function is now called in evens.js. Game now ends when a player has won
or when all spaces on the board are claimed. The cells[] array in
data.js is used to know if all elements have been selected.