Adds local game button functionality
When clicking local multiplayer game button, the game-type selection
hides and the game board appears. A new game is created when the button
is clicked as well, and it is updated on the server while the players
play locally. Player O is always null for a local game and his/her
stats are not tracked.