Changes when getCompletedGames runs
Previously, the GET request for all completed games was made when the
user clicked New Game. Now the function is called when the game-ending
cell is clicked. This makes it so the stats update to reflect the game
that just ended without the user needing to start a new game.